+++ openquake/commands/info.py
