+++ openquake/commands/plot_sites.py
