+++ openquake/commands/purge.py
