+++ openquake/commands/reduce.py
