+++ openquake/commands/run.py
@@ -61,7 +61,7 @@ def get_pstats(pstatfile, n):
     # ====== ======= ========================================================
     # ncalls cumtime path
     # ====== ======= ========================================================
-    # 1      33.502  commonlib/commands/run.py:77(_run)
+    # 1      33.502  commands/run.py:77(_run)
     # 1      33.483  calculators/base.py:110(run)
     # 1      25.166  calculators/classical.py:115(execute)
     # 1      25.104  commonlib/parallel.py:249(apply_reduce)
