+++ openquake/commands/show.py
