+++ openquake/commands/show_attrs.py
